# docs/swagger.yaml
basePath: /
definitions:
  Drom.Car:
    properties:
      carbrand:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.getAllListsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/Drom.Car'
        type: array
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for DromApp Application
  title: Drom App Api
  version: "1.0"
paths:
  /api/lists:
    get:
      consumes:
      - application/json
      description: get all cars from database
      operationId: get-all-cars
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllListsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get All Cars
      tags:
      - cars
    post:
      consumes:
      - application/json
      description: create one Drom to Drom database
      operationId: create-Drom
      parameters:
      - description: Drom info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/Drom.Car'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create Car
      tags:
      - cars
  /api/lists/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Drom by id from database
      operationId: delete-Drom-by-id
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Drom.Car'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Delete Car By Id
      tags:
      - cars
    get:
      consumes:
      - application/json
      description: get Drom by id from database
      operationId: get-Drom-by-id
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Drom.Car'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get Car By Id
      tags:
      - cars
    put:
      consumes:
      - application/json
      description: update Drom by id from database
      operationId: update-Drom-by-id
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: integer
      - description: DromCar
        in: body
        name: Drom
        required: true
        schema:
          $ref: '#/definitions/Drom.Car'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Drom.Car'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Update Car By Id
      tags:
      - cars
swagger: "2.0"